minor improvements in the navbar
