+++ src/data/navbar-data.tsx
@@ -1,0 +1,20 @@
+const links = [
+  {
+    name: "Home",
+    url: "#",
+  },
+  {
+    name: "About",
+    url: "#about",
+  },
+  {
+    name: "Skills",
+    url: "#skills",
+  },
+  {
+    name: "Certifications",
+    url: "#certifications",
+  },
+];
+
+export default links;
